# blog-backend/config/openapi.yaml
openapi: 3.0.3
info:
  title: Gouno Blog API
  description: REST API for managing posts and comments in Gouno Blog.
  version: 1.0.0
servers:
  - url: http://localhost:8080/api
    description: Nginx gateway local entry point
paths:
...
  /posts:
    get:
      summary: List all posts
      description: Retrieve a paginated list of blog posts, optionally filtered by tag.
      parameters:
        - name: tag
          in: query
          description: Filter posts by tag
          required: false
          schema:
            type: string
        - name: page
          in: query
          description: Page number (1-based)
          required: false
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          description: Number of items per page
          required: false
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: A list of posts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostListResponse'
    post:
      summary: Create a new post
      description: Create a new blog post. Requires administrator authentication.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePostRequest'
      responses:
        '201':
          description: Post created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostResponse'
        '400':
          description: Invalid input or missing fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden (Admin role required)
        '409':
          description: Slug already in use
  /posts/{slugOrID}:
    get:
      summary: Get a post by slug or ID
      description: Retrieve a single blog post details by its slug or unique ID.
      parameters:
        - name: slugOrID
          in: path
          description: Unique post ID or url slug
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Post details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostResponse'
        '404':
          description: Post not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /posts/{id}:
    put:
      summary: Update an existing post
      description: Modify an existing blog post details. Requires administrator authentication.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          description: Post numeric ID
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePostRequest'
      responses:
        '200':
          description: Post updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostResponse'
        '400':
          description: Invalid input
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Post not found
        '409':
          description: Slug already in use

    delete:
      summary: Delete a post
      description: Permanently remove a post by ID. Requires administrator authentication.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          description: Post numeric ID
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Post deleted successfully
        '400':
          description: Invalid post ID
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Post not found
...
  /admin/posts:
    get:
      summary: List all posts for administration
      description: Includes draft, scheduled, and published posts. Requires administrator authentication.
      security:
        - BearerAuth: []
      parameters:
        - name: q
          in: query
          schema:
            type: string
          description: Search title, summary, or content
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, scheduled, published]
        - name: category
          in: query
          schema:
            type: string
          description: Category slug
        - name: tag
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 50
            maximum: 100
      responses:
        '200':
          description: Administrative post list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostListResponse'
  /posts/{slugOrID}/comments:
    get:
      summary: List visible comments for a post
      description: Retrieve public comments that an administrator has marked visible under a specific blog post by slug or numeric ID.
      parameters:
        - name: slugOrID
          in: path
          description: Unique post ID or URL slug
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of comments
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentListResponse'
        '400':
          description: Invalid post identifier
        '404':
          description: Post not found
    post:
      summary: Add a comment to a post
      description: Submit a new comment to a specific post by slug or numeric ID. Open to public; newly submitted comments are hidden until an administrator marks them visible.
      parameters:
        - name: slugOrID
          in: path
          description: Unique post ID or URL slug
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCommentRequest'
      responses:
        '201':
          description: Comment created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentResponse'
        '400':
          description: Missing required fields
        '404':
          description: Post not found
  /posts/{slugOrID}/comments/all:
    get:
      summary: List all comments for a post
      description: Retrieve all comments for moderation, including pending hidden comments. Requires administrator authentication.
      security:
        - BearerAuth: []
      parameters:
        - name: slugOrID
          in: path
          description: Unique post ID or URL slug
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of all comments
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentListResponse'
        '400':
          description: Invalid post identifier
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Post not found
  /comments/{id}/visibility:
    put:
      summary: Update comment visibility
      description: Mark a comment visible or hidden for public readers. Requires administrator authentication.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          description: Comment numeric ID
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCommentVisibilityRequest'
      responses:
        '200':
          description: Comment visibility updated
        '400':
          description: Invalid comment ID or request body
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Comment not found
  /comments/{id}:
    delete:
      summary: Delete a comment
      description: Remove a comment by its unique ID. Requires administrator authentication.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          description: Comment numeric ID
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Comment deleted successfully
        '400':
          description: Invalid comment ID
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Comment not found
  /posts/{slugOrID}/community:
    get:
      summary: Get the current visitor's article interaction state
      parameters:
        - name: slugOrID
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Like, bookmark, and current like count
  /posts/{slugOrID}/like:
    put:
      summary: Idempotently like an article
      parameters:
        - name: slugOrID
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Updated interaction state
        '429':
          description: Interaction rate limit exceeded
    delete:
      summary: Remove the current visitor's like
      parameters:
        - name: slugOrID
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Updated interaction state
  /comments/{id}/report:
    post:
      summary: Report a visible comment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
                  maxLength: 500
      responses:
        '201':
          description: Report recorded
        '409':
          description: This visitor already reported the comment
        '429':
          description: Interaction rate limit exceeded
  /me/bookmarks:
    get:
      summary: List the signed-in user's bookmarks
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Bookmarked published articles
        '401':
          description: Unauthorized
  /me/bookmarks/{postID}:
    put:
      summary: Bookmark an article
      security:
        - BearerAuth: []
      parameters:
        - name: postID
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Article bookmarked
    delete:
      summary: Remove an article bookmark
      security:
        - BearerAuth: []
      parameters:
        - name: postID
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Bookmark removed
  /me/notifications:
    get:
      summary: List the signed-in user's notifications and unread count
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Notification page
        '401':
          description: Unauthorized
  /me/notifications/{id}/read:
    put:
      summary: Mark one owned notification as read
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Notification marked read
        '404':
          description: Notification does not exist or belongs to another user
  /me/notifications/read-all:
    put:
      summary: Mark all owned notifications as read
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Notifications marked read
  /admin/comments:
    get:
      summary: List comments across articles for moderation
      security:
        - BearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [all, pending, visible, hidden]
        - name: reported
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: Moderation queue
  /admin/comments/{id}:
    put:
      summary: Change a comment's moderation status
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [pending, visible, hidden]
      responses:
        '200':
          description: Moderation status updated
  /posts/{slugOrID}/related:
    get:
      summary: List related published posts
      parameters:
        - name: slugOrID
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Up to four posts ranked by shared tags
  /admin/posts/{id}/versions:
    get:
      summary: List saved article versions
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: The most recent 50 snapshots
  /admin/posts/{id}/versions/{versionID}/restore:
    post:
      summary: Restore an article version
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: versionID
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Restored article; the replaced content is retained as a snapshot
        '404':
          description: Article version not found
  /admin/media:
    get:
      summary: List media assets
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Media assets ordered newest first
    post:
      summary: Upload an image
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                alt_text:
                  type: string
      responses:
        '201':
          description: Image uploaded
        '400':
          description: Unsupported image type or file larger than 10 MB
  /admin/media/{id}:
    delete:
      summary: Delete a media asset and its local file
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Media deleted
        '409':
          description: Media is still referenced by one or more posts
  /admin/media/{id}/references:
    get:
      summary: List posts that reference a media asset
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Referencing post IDs, titles, and slugs
  /admin/analytics:
    get:
      summary: Get content and engagement analytics
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Aggregate counters, top posts, and a 14-day traffic series
  /admin/provider-profiles:
    get:
      summary: List AI provider profiles
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Provider profiles with masked key metadata
    post:
      summary: Create an AI provider profile
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProviderProfileRequest'
      responses:
        '201':
          description: Provider profile created
        '400':
          description: Invalid provider or upstream URL
  /admin/provider-profiles/{id}:
    put:
      summary: Update an AI provider profile
      description: Omit api_key to retain the existing encrypted secret.
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ResourceID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProviderProfileRequest'
      responses:
        '200':
          description: Provider profile updated
    delete:
      summary: Delete an unused AI provider profile
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ResourceID'
      responses:
        '200':
          description: Provider profile deleted
        '409':
          description: Provider profile is used by an Agent
  /admin/provider-profiles/{id}/test:
    post:
      summary: Test an AI provider connection
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ResourceID'
      responses:
        '200':
          description: Connection succeeded and latency was recorded
  /admin/agents:
    get:
      summary: List AI Agents
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Configured Agents
    post:
      summary: Create an AI Agent
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentRequest'
      responses:
        '201':
          description: Agent created
  /admin/agents/{id}:
    get:
      summary: Get one AI Agent
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ResourceID'
      responses:
        '200':
          description: Agent configuration
    put:
      summary: Update an AI Agent
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ResourceID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentRequest'
      responses:
        '200':
          description: Agent updated
    delete:
      summary: Soft-delete an AI Agent
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ResourceID'
      responses:
        '200':
          description: Agent deleted
  /admin/agents/{id}/enable:
    post:
      summary: Enable an AI Agent
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ResourceID'
      responses:
        '200':
          description: Agent enabled and next run calculated
  /admin/agents/{id}/disable:
    post:
      summary: Disable an AI Agent
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ResourceID'
      responses:
        '200':
          description: Agent disabled
  /admin/agents/{id}/run:
    post:
      summary: Queue an immediate AI Agent run
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ResourceID'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                input:
                  type: object
                  additionalProperties: true
      responses:
        '202':
          description: Run queued
        '409':
          description: Run limit, token budget, or active-run conflict
  /admin/agent-tools:
    get:
      summary: List Blog tools available to AI Agents
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Tool schemas and risk levels
  /admin/agent-presets:
    get:
      summary: List built-in AI Agent templates
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Weekly operations, content health, and comment insight templates
  /admin/agent-runs:
    get:
      summary: List AI Agent runs
      security:
        - BearerAuth: []
      parameters:
        - name: agent_id
          in: query
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 50
            maximum: 100
      responses:
        '200':
          description: Paginated run history
  /admin/agent-runs/{id}:
    get:
      summary: Get an AI Agent run and its tool-call timeline
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ResourceID'
      responses:
        '200':
          description: Run and tool calls
  /admin/agent-approvals:
    get:
      summary: List AI Agent approvals
      security:
        - BearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            default: pending
            enum: [pending, approved, rejected, expired, executed, failed]
      responses:
        '200':
          description: Paginated approval queue
  /admin/agent-approvals/{id}/approve:
    post:
      summary: Approve and execute an AI Agent proposal
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ResourceID'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApprovalReviewRequest'
      responses:
        '200':
          description: Proposal approved and executed
        '409':
          description: Proposal expired or target changed after proposal
  /admin/agent-approvals/{id}/reject:
    post:
      summary: Reject an AI Agent proposal
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ResourceID'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApprovalReviewRequest'
      responses:
        '200':
          description: Proposal rejected
...
  /categories/{slug}/posts:
    get:
      summary: List published posts in a category
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: Paginated category posts
        '404':
          description: Category not found
...
  /admin/categories:
    get:
      summary: List categories for administration
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Category list
    post:
      summary: Create a category
      security:
        - BearerAuth: []
      responses:
        '201':
          description: Category created
  /admin/categories/{id}:
    put:
      summary: Update a category
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Category updated
    delete:
      summary: Delete a category and unclassify its posts
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Category deleted
  /admin/tags:
    get:
      summary: List tags with post counts
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Tag summaries
  /admin/settings:
    get:
      summary: Get editable site settings
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Site settings
    put:
      summary: Update allow-listed site settings
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Updated site settings
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    PostResponse:
      type: object
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: "success"
        data:
          $ref: '#/components/schemas/Post'
    PostListResponse:
      type: object
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: "success"
        data:
          type: object
          properties:
            list:
              type: array
              items:
                $ref: '#/components/schemas/Post'
            total:
              type: integer
              example: 45
            page:
              type: integer
              example: 1
            pageSize:
              type: integer
              example: 10
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: integer
          example: 400
        message:
          type: string
          example: "Key: 'CreatePostRequest.Title' Error:Field validation for 'Title' failed on the 'required' tag"
        data:
          type: object
          nullable: true
          example: null